requirements.txt
fastapi==0.104.1
praw==7.7.1
chromadb==0.5.0
anthropic>=0.45.2
openai>=1.13.0
pydantic==2.5.0
python-dotenv==1.0.0
pytest==7.4.3
structlog==23.2.0
httpx==0.25.2
uvicorn==0.24.0
gradio==4.36.0
numpy<2.0.0
